# repair the integrated branches against each other
Collisions that only appear once the branches sit side by side; no behaviour
change intended.

- Two branches each added an IArchiveModifiable overload, making every bare
  <see cref="IArchiveModifiable.Add"/> and .Remove ambiguous (CS0419).
- UefiFvReader and four fixtures captured a ReadOnlySpan in a lambda.
- NrgWriter allocated inside a loop (CA2014).
- NRG reached RebuildVerb.CreateScratchStream from another assembly, where it
  was internal.
- An external-oracle fixture imported System.ComponentModel for Win32Exception,
  making [Category] ambiguous against NUnit's.
- CSO fixtures called a default interface member on the concrete descriptor.
- A Lustre fixture's Math.Max(int, ushort) matched several overloads.
- Two fixtures built fixed-capacity MemoryStreams for verbs that grow the image.

# the shared maintenance probe encrypted archives the verbs cannot open

Making the probe credential-aware meant every password-capable format got an
encrypted probe -- but Defragment(Stream) and Purge take no credentials, so
Zip, 7z, Rar, Ace, Arj, Sqx, Cb7 and Cbr all failed their round trips. The probe
is unencrypted, falling back to a password only where creation refuses without
one.

* regenerate the filesystem support matrix from the live registry

+++ Compression.Tests/Lustre/LustreDetectionTests.cs
@@ -124,7 +124,7 @@ public class LustreDetectionTests {
     if ((featureIncompat & 0x80) != 0)
       totalBlocks |= (ulong)BinaryPrimitives.ReadUInt32LittleEndian(sb.Slice(0x150, 4)) << 32;
     var descriptorSize = (featureIncompat & 0x80) != 0
-      ? Math.Max(32, BinaryPrimitives.ReadUInt16LittleEndian(sb.Slice(0xFE, 2)))
+      ? Math.Max(32, (int)BinaryPrimitives.ReadUInt16LittleEndian(sb.Slice(0xFE, 2)))
       : 32;
     var groupCount = (totalBlocks - firstDataBlock + blocksPerGroup - 1) / blocksPerGroup;
     var descriptorTableOffset = (long)(firstDataBlock + 1UL) * blockSize;
